[Refactor] create config/
