+++ src/main/java/yeonjeans/saera/config/AppConfig.java
@@ -1,4 +1,4 @@
-package yeonjeans.saera;
+package yeonjeans.saera.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
